feat: split scrapeData into sub tasks (#209)
* feat: move fetching scrapers to its own task

* feat: break up scrapeData

* Add reporting back

* Remove options

* address code review

* Update tasks/scrapeData/runScraper.js


* Update lib/transform.js


* Update tasks/scrapeData/runScraper.js


* code review
