+++ cli.js
@@ -1,6 +1,8 @@
 const imports = require('esm')(module);
 
-const generate = imports('./index.js').default;
+const generate = imports('./tasks/index.js').default;
 const argv = imports('./lib/cliArgs.js').default;
 
-generate(argv.date, argv);
+generate(argv.date, argv).catch(e => {
+  throw e;
+});
