+++ lib/transform.js
@@ -1,6 +1,28 @@
 import crypto from 'crypto';
 import * as datetime from './datetime.js';
 
+const scraperVars = ['type', 'timeseries', 'headless', 'ssl', 'priority', 'aggregate'];
+
+/*
+  Remove "private" object properties
+*/
+export const removePrivate = data => {
+  for (const [prop, value] of Object.entries(data)) {
+    if (value === '' || prop[0] === '_') {
+      delete data[prop];
+    }
+  }
+
+  return data;
+};
+
+export const removeScraperVars = location => {
+  // Remove non-data vars
+  for (const prop of scraperVars) {
+    delete location[prop];
+  }
+};
+
 /*
   Convert an object keyed on county name to an array
 */
